+++ src/importer/webanno.rs
@@ -13,10 +13,7 @@ use pest::{Parser, iterators::Pair};
 use pest_derive::Parser;
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    importer::GenericImportConfiguration, progress::ProgressReporter,
-    util::graphupdate::import_corpus_graph_from_files,
-};
+use crate::{importer::GenericImportConfiguration, progress::ProgressReporter};
 
 use super::Importer;
 
@@ -36,8 +33,7 @@ impl Importer for ImportWebAnnoTSV {
         tx: Option<crate::workflow::StatusSender>,
     ) -> Result<graphannis::update::GraphUpdate, Box<dyn std::error::Error>> {
         let mut update = GraphUpdate::default();
-        let paths_and_node_names =
-            import_corpus_graph_from_files(&mut update, input_path, &config)?;
+        let paths_and_node_names = config.derive_corpus_graph(input_path, &mut update)?;
         let progress =
             ProgressReporter::new(tx.clone(), step_id.clone(), paths_and_node_names.len())?;
         for (pathbuf, doc_node_name) in paths_and_node_names {